chore: 기능별로 js 파일 분리하고 폴더 정리하기
