--- week2/assignment/data.js
@@ -1,296 +1,0 @@
-export const members = [
-  {
-    id: 1,
-    name: "강민하",
-    englishName: "MinHaKang",
-    github: "m2na7",
-    gender: "male",
-    role: "YB",
-    firstWeekGroup: 6,
-    secondWeekGroup: 4,
-  },
-  {
-    id: 2,
-    name: "공준혁",
-    englishName: "JunHyukKong",
-    github: "ocahs9",
-    gender: "male",
-    role: "OB",
-    firstWeekGroup: 5,
-    secondWeekGroup: 7,
-  },
-  {
-    id: 3,
-    name: "곽지욱",
-    englishName: "JiUgGwag",
-    github: "gwagjiug",
-    gender: "male",
-    role: "YB",
-    firstWeekGroup: 9,
-    secondWeekGroup: 1,
-  },
-  {
-    id: 4,
-    name: "김가연",
-    englishName: "GaYeonKim",
-    github: "ayla_12",
-    gender: "female",
-    role: "OB",
-    firstWeekGroup: 8,
-    secondWeekGroup: 7,
-  },
-  {
-    id: 5,
-    name: "김가현",
-    englishName: "GaHyeonKim",
-    github: "maylh",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 4,
-    secondWeekGroup: 8,
-  },
-  {
-    id: 6,
-    name: "김건휘",
-    englishName: "GeonHwiKim",
-    github: "KIMGEONHWI",
-    gender: "male",
-    role: "OB",
-    firstWeekGroup: 2,
-    secondWeekGroup: 8,
-  },
-  {
-    id: 7,
-    name: "김고은",
-    englishName: "GoEunKim",
-    github: "gonn-i",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 2,
-    secondWeekGroup: 2,
-  },
-  {
-    id: 8,
-    name: "김다현",
-    englishName: "DaHyunKim",
-    github: "daahyunk",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 5,
-    secondWeekGroup: 9,
-  },
-  {
-    id: 9,
-    name: "김린아",
-    englishName: "RinAKim",
-    github: "Rinakk",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 2,
-    secondWeekGroup: 9,
-  },
-  {
-    id: 10,
-    name: "김민정",
-    englishName: "MinJeongKim",
-    github: "minjeoong",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 4,
-    secondWeekGroup: 7,
-  },
-  {
-    id: 11,
-    name: "김채은",
-    englishName: "ChaeEunKim",
-    github: "bongtta",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 1,
-    secondWeekGroup: 5,
-  },
-  {
-    id: 12,
-    name: "김채현",
-    englishName: "ChaeHyunKim",
-    github: "imddoy",
-    gender: "female",
-    role: "OB",
-    firstWeekGroup: 8,
-    secondWeekGroup: 1,
-  },
-  {
-    id: 13,
-    name: "김태욱",
-    englishName: "TaeWookKim",
-    github: "Taew00k",
-    gender: "male",
-    role: "YB",
-    firstWeekGroup: 8,
-    secondWeekGroup: 6,
-  },
-  {
-    id: 14,
-    name: "김한서",
-    englishName: "HanSeoKim",
-    github: "seueooo",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 6,
-    secondWeekGroup: 3,
-  },
-  {
-    id: 15,
-    name: "노찬영",
-    englishName: "ChanYoungNoh",
-    github: "shroqkf",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 9,
-    secondWeekGroup: 2,
-  },
-  {
-    id: 16,
-    name: "문성희",
-    englishName: "SeonghuiMoon",
-    github: "seong-hui",
-    gender: "female",
-    role: "OB",
-    firstWeekGroup: 9,
-    secondWeekGroup: 4,
-  },
-  {
-    id: 17,
-    name: "박채연",
-    englishName: "ChaeYeonPark",
-    github: "chaeneey",
-    gender: "female",
-    role: "OB",
-    firstWeekGroup: 1,
-    secondWeekGroup: 9,
-  },
-  {
-    id: 18,
-    name: "박희선",
-    englishName: "HeeSunPark",
-    github: "heesunee",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 3,
-    secondWeekGroup: 3,
-  },
-  {
-    id: 19,
-    name: "배영경",
-    englishName: "YoungKyungBae",
-    github: "bykbyk0401",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 6,
-    secondWeekGroup: 2,
-  },
-  {
-    id: 20,
-    name: "손지유",
-    englishName: "JiYooSon",
-    github: "zzz-myam",
-    gender: "female",
-    role: "OB",
-    firstWeekGroup: 9,
-    secondWeekGroup: 5,
-  },
-  {
-    id: 21,
-    name: "유서연",
-    englishName: "SeoYeonYu",
-    github: "youtheyeon",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 3,
-    secondWeekGroup: 1,
-  },
-  {
-    id: 22,
-    name: "유태승",
-    englishName: "TaeSeungYoo",
-    github: "gudusol",
-    gender: "male",
-    role: "OB",
-    firstWeekGroup: 2,
-    secondWeekGroup: 2,
-  },
-  {
-    id: 23,
-    name: "이예림",
-    englishName: "YeRimLee",
-    github: "yarimu",
-    gender: "female",
-    role: "OB",
-    firstWeekGroup: 4,
-    secondWeekGroup: 6,
-  },
-  {
-    id: 24,
-    name: "이윤지",
-    englishName: "YoonJiLee",
-    github: "Leeyoonji23",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 5,
-    secondWeekGroup: 4,
-  },
-  {
-    id: 25,
-    name: "이진혁",
-    englishName: "JinHyeokLee",
-    github: "constantly-dev",
-    gender: "male",
-    role: "YB",
-    firstWeekGroup: 1,
-    secondWeekGroup: 3,
-  },
-  {
-    id: 26,
-    name: "임화랑",
-    englishName: "HwaRangIm",
-    github: "thisishwarang",
-    gender: "male",
-    role: "OB",
-    firstWeekGroup: 3,
-    secondWeekGroup: 5,
-  },
-  {
-    id: 27,
-    name: "최민",
-    englishName: "MinChoi",
-    github: "Minn-Choi",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 1,
-    secondWeekGroup: 6,
-  },
-  {
-    id: 28,
-    name: "최정완",
-    englishName: "JungWoanChoi",
-    github: "look-back-luca",
-    gender: "male",
-    role: "YB",
-    firstWeekGroup: 4,
-    secondWeekGroup: 9,
-  },
-  {
-    id: 29,
-    name: "한수정",
-    englishName: "SooJeongHan",
-    github: "hansoojeongsj",
-    gender: "female",
-    role: "YB",
-    firstWeekGroup: 8,
-    secondWeekGroup: 8,
-  },
-];
-
-if (!localStorage.getItem("membersData")) {
-  localStorage.setItem("membersData", JSON.stringify(members));
-}
