+++ week2/assignment/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="./style/style.css" />
     <title>관리자페이지</title>
   </head>
   <body>
@@ -191,6 +191,6 @@
       src="https://kit.fontawesome.com/11cad02120.js"
       crossorigin="anonymous"
     ></script>
-    <script type="module" src="./index.js"></script>
+    <script type="module" src="./memberList.js"></script>
   </body>
 </html>
