+++ week2/assignment/style/style.css
@@ -40,6 +40,7 @@
   width: 100%;
   background-color: var(--lightgray);
   padding-top: 2rem;
+  padding-bottom: 5rem;
 }
 
 /* member-filter */
